In-game Pause Menu
Added an In-game pause menu, that works by just pressing "ESC".

The Resume and Quit button work, though the Quit button needs to be hooked up to the build of the game.

The Menu button doesn't work, as it's going to be changed to "Option" button further down the line.